+++ lib/bloc_morph.dart
@@ -1,2 +1,4 @@
-export 'core/bloc_morph_gen.dart';
-export 'core/morph_state.dart';
+export 'src/bloc_morph_gen.dart';
+export 'src/morph_state.dart';
+export 'src/morph_selector_multiple.dart';
+export 'src/morph_selector.dart';