+++ lib/src/bloc_morph_gen.dart
@@ -1,4 +1,4 @@
-import 'package:bloc_morph/core/morph_state.dart';
+import 'package:bloc_morph/src/morph_state.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';